Explicitly reintroduced button palette mapping
As part of establishing `_foundationSettings` as an override file, I carefully evaluated all of the variables that had been added, changed, or removed since we had originally forked that file several years ago (79c66e8). As part of that, I had removed a change to how `$button-palette` was mapped, since it should inherit that value as a default (d4c2892).

Unfortunately, this doesn't seem to be the case. As discovered with `$global-flexbox`, some variables don't have their defaults (correctly) implemented, and thus need to be explicitly added to the settings files (91a2615). This appears to also be the case with `$button-palette`. Without it being defined, most buttons work fine, but the white buttons on the `Home.cshtml` are rendered as yellow.

This isn't the most elegant fix, but this is remedied by explicitly setting the default value of `$button-palette` in `_foundationSettings`, as it includes an explicit mapping for `$palette-white`.