Updated gulp to compile all entry points, regardless of location
Previously, the `gulpFile.js` was configured to explicitly compile the `Style.scss` and any styles under `/Views`. It's now modified to simply publish any `*.scss` file which doesn't begin with an underscore, regardless of location. This is much more flexible, and allows us to introduce, if we choose, other locations for stand-alone files—such as `/Components`, if that makes sense.