Replaced `FileIcons.js` with CSS-based application
Previously, `FileIcons.js` evaluated every link on the page and used jQuery to evaluate the page for links ending with a particular file type extension, and then dynamically added a Font Awesome icon after matching links. While that process is surprisingly fast, it's still complicated and adds an extra step for the JavaScript to evaluate on page load.

Since we've since integrated Font Awesome with our SASS structure, we can instead handle this exclusively in CSS. There are some limitations—for instance, the underline extends under the icon. But it's considerably simpler, and likely faster. (Though, on the last point, I haven't measured the difference.)