Replace import of `fontawesome` with just `_mixins`, `_variables`
Previously, CKEditor had the entirety of Font Awesome being imported, which added considerable weight to the already weighy `CKEditor.css` file. Ironically, the icons didn't work inside the editor since I'm not importing the font (oops). But, more over, this wasn't even needed since we're not (currently) expecting GoldSim's editors to be defining Font Awesome fonts inline, and the styles that depend upon them are using the inline format anyway.

As such, I've removed the `@import` of `fontawesome` and instead individually included `_mixins`, `_variables`, and the `regular` font. This reduces the stylesheet from 76k to 21k. A considerable improvement!