Moved `_accordion` partial to `Accordion` view-specific stylesheet
Between the Foundation Sites' `foundation-accordion` import, and all of GoldSim's own overrides, the accordion stylesheet ends up being 9k. But only a few pages even use it—and they're not among GoldSim's most popularly referenced pages. As such, by migrating this to a component-specific stylesheet, we're able to reduce the global stylesheet by a further 9k, while deferring this until users hit e.g. Technical Papers, Application Index, Application Page, or, of course, Accordion.

This also represents the first time we've used the newly introduced ability to create component-specific stand-alone stylesheets (cc41460). To help establish a best practice here, I left the reference in `_components.scss`, but commented it out—that way it's clear this isn't an oversight, and helps document that if an implementer needs a reference to the accordion then they'll need to include that stand-alone stylesheet. We may revisit this later, but as an initial structure this seems effective.

(Another option would be to establish a separate directory for stand-alone component style sheets, e.g. under `/Views`, but I'm not yet convinced that's the most intuitive approach.)