Removed empty headings in SASS files
While there are sometimes good reasons to have empty comment headers in files as a placeholder, most of these will never be used—and can always be (re)introduced should the need arise. There are a few of these I left in for cases where they make sense (e.g., as a placeholder for global styles).