Removed unnecessary and redundant SASS `@import`s
A number of stylesheets `@import`ed `_variables.scss` and/or `_foundationSettings.scss`. But these are included in the base `Style.scss` and aren't (strictly) needed in each individual stylesheet—at least as far as the SASS compiler is concerned (IDE tooling might be another matter). Regardless, they're really inconsistently applied as is, so I'm removing them for now.