Namespaced Font Awesome path and centralized `@import`
In a previous commit, I added an `includePath` for Font Awesome to the SASS compilation in the `gulpFile.js`, thus allowing us to shorten the cumbersome relative path references. Unfortunately, that introduced unintuitively scoped includes—like `_mixins.scss` and `_variables.scss`. While these weren't (yet) ambiguous references as far as the compiler was concerned, they definitely introduced ambiguity from the maintainers' perspective.

To mitigate this, I changed the `includePath` to the `@fortawesome` container, thus requiring that Font Awesome includes be prefixed with `fontawesome/scss`. That reintroduces some length to the `@import` paths, it is true, but in doing so forces the references to be namespaced, thus eliminating the (potential) ambiguity.

While I was at it, I also moved the includes of Font Awesome's `_mixins` and `_variables` from `_icons` to the main `Style.scss` file, since we're now _exclusively_ using those as our way of defining Font Awesome icons, and have incorporated them across a variety of locations (f7cfb97, c0afdc5, 0140ed4, 0f1b12d). They still need to be included in stand-alone files, but it makes sense for the Font Awesome mixins and variables to be available within any partials, not just `_icons`.

I maintained the `includePath` for `foundation-sites`. That's no longer strictly needed since we no longer include `settings` (which had a root relative reference to `util/util`). But since most of Foundation's partials are uniquely named, there isn't as strong of a need for namespacing, and keeping it is recommended by Zurb when compiling our own SCSS since some of their partials DO include root relative paths.

That said, if we encounter ambiguous Foundation references in the future, we might revisit this by adding a path to `node_modules` and importing (at least) those ambiguous references with a `foundation-sites/scss/` prefix.