Consolidated and reorganized Page Level Navigation styles
For reasons that are unclear, the Page Level Navigation was previously distributed among `_base` and `_menu`.  I identified the former previously, and moved it into a new `_pageLevelNavigation` component partial, which makes more sense (365df66). I had missed, however, that the navigation styles themselves were part of `_menu`. This commit corrects that by moving those styles into the previously established `_pageLevelNavigation`.

While I was at it, I also reorganized the headers for the `_pageLevelNavigation` to be more specific and descriptive.