Refactored slideshow script as jQuery plugin
Moved the OwlCarousel initialization script out of the `Slideshow.cshtml` template and into a new `Slideshow.js` JavaScript file. As part of this, wrapped it in a (very basic) jQuery plugin so that it can be easily initialized.

This likely isn't following best practices for a jQuery plugin. There isn't a configuration. It presumes that Owl Carousel (`Owl`) is already loaded without establishing it as a dependency. It currently hard-codes values regarding the paging classes.

As this is not intended to be a reusable library, but simply a wrapper for configuring Owl carousel, the above limitations are acceptable. In fact, we're primarily using the jQuery plugin model because it provides a familiar and intuitive instantiation, and helps facilitate packaging the code in a consistent format that's easy to read and document.

This is similar to the approach we previously took with `Search.js` (3409dfe), though that exposes more control over the nested structure.