Established new `InlineScript` block and moved all scripts to it
Previously, all script references as well as inline scripts defined in the page were placed within a `@section Script` block.

This worked. However, most inline scripts should be wrapped in a check for the ready state, and not all were. Instead of updating each of these script blocks to do that, I've instead established a new `InlineScript` section which is wrapped in a listener for `DOMContentLoaded`.

This not only separates out inline scripts from script references, but helps enforce that inline page scripts don't run until the DOM is ready. Since most inline scripts are wiring up jQuery events, this is especially useful.

That said, it's worth noting that this doesn't use the jQuery `$(document).ready(…)` handler, instead preferring the non-jQuery `window.addEventListener('DOMContentLoaded', …)` handler. These are nearly equivalent functionally, but the new approach, while a bit wordier, allows this to run while waiting for jQuery to load. As a result, we'll be able to defer the loading of jQuery in a future commit.

For scripts that were already (correctly) nested inside of a `$(document).ready(…)` block, this change has resulted in those scripts being unintented, which makes this change appear bigger than it is. Those scripts have not changed, they've just been pulled back four spaces. This is especially notable on the forms, which have pretty large inline code blocks (Perhaps these should be refactored into stand-alone scripts, but that's a separate task that we'll need to consider later.)