Imports Foundation scripts individually, instead of as a bundle
Previously, we simply included `foundation.min.js` in the `_layout.cshtml`. That's an easy approach, but it yields a larger payload that we require, since we only use about half of Foundation's JavaScript modules.

To remedy that, I've broken this down into individual script calls. This adds a lot of clutter to the `_layout.cshtml` template, but reduces the total payload from 178k to 124k—a 30% reduction.

There is an argument that these should be bundled, which would reduce the clutter and allow them to be downloaded as a single package. This argument isn't as strong now that we use HTTP/2, however, as now these can all be downloaded in parallel without blocking eachother or other resources. In fact, it may well be that this approach is _faster_ than a bundle due to that parallelism—though there are tradeoffs in terms of gzip performance.

That said, we'll likely reevaluate the bunding in the future. The reason we didn't do this now is that bundling Foundation correctly requires the use of an ES6 transpiler, such as Babel. Once we migrate our scripts to ES6, and integrate e.g. Babel into our Gulp workflow, we'll revisit this. Until then, we could alternatively bundle the transpiled plugins into a single script—but, again, it's debatable whether that gains us anything in terms of performance at this point.

It's worth noting that the most cumbersome part of this is managing the plugins. Each JavaScript plugin has dependencies on one or more plugins, so these need to be loaded upfront. If and when we migrate to an ES6 based transpilation, I believe this will be simpler—either because the imports will be inherited from parent modules, or because tree shaking will automatically clean out unused utilities.

Finally, since the accordion script is only used on a handful of pages, that script is placed in the individual templates that need it, just like our own `ScrollFix.js`. This allows us to defer the loading of that until it's actually required. The rest are potentially needed on any page—with most being mandated by the `_layout.cshtml` itself—and are thus centralized.