Marked all JavaScript references as `defer`
JavaScript files aren't normally needed until after the page has rendered. In fact, assumng they operate against the DOM, they specifically CAN'T be loaded until after the page has rendered. As such, they can—and should—be marked as `defer`. This instructs the browser to deprioritize loading them behind other resources, so that they're not blocking those resources. This isn't as big of an issue with HTTP/2 since there isn't a limit on the number of resources that can be downloaded, but it still helps the browser prioritize available bandwidth, should that be a concern.

This is made possible by the establishment of the new `InlineScript` section (efd1931), which ensures that inline scripts can be initially executed even though jQuery isn't yet loaded. Without that, they would have tripped over the fact that `$` isn't available when executing `$(document).ready(…)`. (This is an issue since inline scripts can't be deferred.)

While I was at it, I also moved the `defer` keyword to the beginning of the script to ensure it's in a consistent location. This makes it clear that they're deferred, without needing to scan different locations in the script tag to confirm. The Foundation plugins setup in the last commit (5705a4c) already followed this format.