Setup DASH to only import `mediaplayer`
Previously, DASH was setup to import the full 589k DASH.js script (which is huge!). That includes a number of plugins for e.g. offline playback, DRM, and reporting that we're not currently using. Given that, I've modified this to instead import just the `mediaplayer` component. That's _still_ huge at 518k, but it at least trips off 70k—or 12%.

**Note:** I prematurely committed the reference to this in `Home.cshtml` as part of the implementation of `defer` (569509f). I do apologize, future reader, for the sloppy commit history. Ugh.