Update _Layout.cshtml
Last week, I moved Application Insights to use a locally defined script, which permitted us to cache the JavaScript file and significantly reduce the inline script for initializing the package (a4eef2d). When I did that, however, I neglected to remove the _original_ inline implementation of Application Insights! As such, it was being downloaded and initialized twice. Oops!

This fixed that issue by removing the (now) redundant section. Note that when the new format was introduced, it was also moved to the `head` section, per Microsoft's guidance. That way, any errors that occur on the page will be properly logged by Application Insights. (It's the same reason the Google Analytics script is in the `head`, even though scripts are usually loaded at the end of the `body`.)