Removed Image Zoom capabilities
The original design allowed for a large image to be embedded in a page, and then annotated with a "Zoom In" button which would open it in a full-size modal window. While the implementation was pretty neat, there ended up not being a need for it on the site, and thus it was never actually used. Given that, I'm removing the JavaScript and style references associated with it.