Ensured all jQuery functions are wrapped in a `ready()` event
While our global scripts were all wrapped in a `$(document).ready(…)` event, most of the local page scripts were not. This didn't seem to be causing a problem on any of the pages, but it certainly introduces the potential to. While this is far less likely now that we've deferred these scripts—and, thus, they won't usually be loaded and evaluated until after the DOM is ready—it's still a best practice we should abide by to be safe.

Note that this results in all jQuery wireups being intended by two spaces, which makes this update appear more invasive than it actually is.