Fixed references to previous page button in search script
For reasons that are unclear, the next button references were consistencly referenced as `_nextButton`, but the previous button references were all over the board: `backButton` and `previousPage` were used. This was likely due to an improperly implemented (and tested) refactoring of the variables. This is fixed now by consistently naming them `previousButton`.