Removed support for full-height panes
Originally, the first pane on the homepage occupied the entire height of the window (minus the top header). This was a nice effect in that it created a more immersive experience. That said, it ended up being removed fairly early on as there were concenrs about how the content filled out the space. Personally, I'd like to see it reintroduced, but in its absence there's no benefit to maintaining the CSS or, more importantly, the jQuery script which looks for `js-full-height` panes on _every_ page. If there's more support for reintroducing this later, then we can revisit it then.