Fixed Awesome Foundation Checkboxes implementation
There were two significant issues with the implementation of Awesome Foundation Checkboxes, both of which date back to the upgrade to .NET Core in December.

The first is that I was, apparently, no longer importing the stylesheet, despite having setup `package.json` to download the dependencies via npm. That is easily fixed by simply `@import`ing the source code directly into our (new) `Form.scss` stylessheet.

The second is because, it seems, Font Awesome no longer supports that checkbox icon in the current version of the regular font, so when we upgraded that broke. Font Awesome removed a number of common, basic icons from the regular set, which has introduced several issues like this. Fortunately, this is easily resolved by explicitly setting the `font-weight` to 700, which instructs it to load the icon from the solid font instead of the regular font. The checkbox still exists in the solid font.

While I was at it, I added the `node_modules` folder to the `includePaths` for the SCSS compilation. This allows us to reference source stylesheets from npm packages without needing to add each individual path to our `includePaths`, as I previously did for `foundation-sites` and `@fortawesome`. As we reference Foundation and Font Awesome with some frequency, it's handy having those two setup explicitly, but for other items which we'll likely only call once, such as `awesome-foundation6-checkbox`, having a generic root still allows us to use root-relative paths without needing to explicitly include each and every package we want to reference here. It also ensures those paths are "namespaced" so that we're avoiding potential ambiguities over `@import`s (as would have likely happened here due to `settings.scss` in the `awesome-foundation6-checkbox` source folder.