Update Home.scss
In a previous commit, I moved the padding on the Application cards from the links themselves to the newly reintroduced `card-section` (93e5b1d). That seemed like a good idea in that it shared the padding between both the list as well as the title. In doing so, however, it prevented the hover state from bleeding to the edge, which looks goofy. Oops.

This commit undoes that change, and extends the padding to the header, thus acheiving the original objective while maintaining the hover bleed.