Fixed Owl Carousel navigation icons on the homepage
When I refactored Font Awesome to exclusively rely on the Font Awesome mixins (ff8e51e), instead of outputting classes to the CSS which we may or may not use, I neglected to account for the Owl Carousel implementation on the home page, which configures the icons via script. (It seems I was only looking in CSHTML files for implementations of Font Awesome.)

This is easily fixed by moving these to the CSS using the Font Awesome mixins. While I was at it, I refactored the markup and class names to better reflect that these aren't standard Font Awesome classes. Welcome back, little ones!