Repositioned icons on new styles
With the courseware, we introduced two new styles into the stylesheet: `blockquote.note` and `aside.help`. To keep these easy to maintain, the iconography is defined inside the CSS using the Font Awesome mixins. This works well, but when I first created this I was unable to change the positioning, which looked a bit goofy in some scenarios, and concerned me that this approach didn't offer flexibility.

@ktrunkey identified a solution, however, which gives us much more control. This commit applies that fix.

While I was at it, I fine-tuned the margins so they better integrated with the overall layout, and with one another. Now, both styles are indented by `rem-calc(44)`, with the icon being flush with the left side. This looks much nicer than my original implementation.