Implemented alternate styling for aligning lesson navigation
In a previous commit, I applied some CSS using `margin-left` in order to align the text in the lesson navigation so that it cleared the progress tracking icon when wrapping (c9addb0). This approach worked, but was really unintuitive and touchy—the arithmetic doesn't align with anything obvious, and small changes in the numbers can have a large impact on the output.

@ktrunkey identified an alternate approach using `position: absolute` that allows us to move away from `text-indent`. The arithmetic isn't perfect—for instance, we need to nudge the icon down based on an assumption of what the calculated `height` will end up being. But it's much more predictable and intuitive than the original implementation.

As part of this, I was also able to tighten up the margins a bit to maximize the space for the text without making it feel to compressed. I had wanted to do this previously, but it was proving difficult to accomplish with the unpredictability of the `text-indent` implementation.