Estasblished `preconnect` and `dns-prefetch` hints
The `preconnect` hints are for servers that _will_ be touched (assuming data isn't cached) _and_ aren't part of the top-line metadata (i.e., referenced directly by `src` elements in the `<head>`). This is especially useful for the secondary services called by Google Analytics and Application Insights, as those will always be hit, but the actual URLs are buried behind an initial service call (e.g., `www.google-analytics.com` triggers a call to `stats.g.doubleclick.net`).

The `dns-prefetch` is a more tentative lookup. It doesn't establish a connection, as it's not known whether there will be content images on the page. But it at least ensures that the browser is kicking off the DNS lookup first, so that the first call to any such resources can rely upon that.

In practice, it's unclear how much of a difference this will make for the browsers which support it. It's likely a marginal impact. My initial performance testing doesn't suggest a major gain in e.g. Chrome—but it's also not demonstrating any real penalty, either. Given that, we may want to reevaluate these over time. For now, however, I'm following these per Google's best practices.