Optimized splash screen by screen resolution, browser capabilities
The splash screen background is big. _Really_ big. Like 373k big. Ouch.

To better optimize this, I created smaller versions at 1920px, 1366px, 1024px, and 640px to align both with our breakpoints, but also with popular screen resolutions used by GoldSim customers. In addition, I added a check for WebP support and, if the browser supports it, point instead to a WebP file.

This process is a bit ugly. If we want to reproduce this for other images (on the homepage, or elsewhere) we will want to make a mixin for this—if not something in our build process that automatically generates different size images automatically. For this single image, though, that's not yet necessary.

As this produced a lot of binary noise, I also moved the images off to `media`. We should avoid storing binaries in git, but especially here where it's just a repetition of the same image in different sizes, formats.