Removed top offset for closed icon in menu
For some reason, the close ("-") icon was being pushed down by 0.25 rems (`rem-calc(4)`). This might have made sense given previous defaults, possibly in Foundation Sites. Currently, though, it means the "-" ends up aligned with the bottom of the text. By removing this, it inherits the `top: 0` attribute, thus aligning it with the open ("+") button while it's in the closed state.