Updated selector for menu item padding
The menu item padding was being based on `li[role="menuitem"]`. There are two issues with this.

First, the Aria role is assigned dynamically by Foundation, meaning that the padding doesn't get applied until after Foundation has initialized the Dropdown Menu. This causes an annoying flash of unstyled content (FOUC).

Second, in Foundation 6.6.2, the `menuitem` role is no longer assigned to menu items. It's unclear exactly why this is since they are, in fact, menu items. But, regardless, this selector no longer functions in the latest patches. To ensure we can keep up-to-date with Foundation, this needs to be updated.