Added `Replace()` action
The `Replace` action (`/Administration/Utility/Replace`) accepts two query string parameters (`searchText` and `replaceText`) and will find all topics containing the `searchText` and replace it with the `replaceText`. Obviously, this should be used with great caution! The `Find()` report (6d8c3de) can be used to help validate the scope.