Added `UpdatePaths` action
The `UpdatePaths()` action (`/Administration/Utility/UpdatePaths`) finds all local references to content items (either via absolute or relative paths) and updates them to point to the new `media` server.

This utility should only need to be run once. It's being maintained, however, as it provides a useful boilerplate for supporting similar requests in the future, while simultaneously being difficult to generalize for those purposes. As such, it should provide a useful reference.