Centralize `_ViewStart` and `_ViewImports`
The `_ViewStart.cshtml` and `_ViewImports.cshtml` were originally in their default locations under `/Views`. This meant tht the settings established in each needed to be reintroduced for each of the individual areas, since the `/Areas` folder doesn't inherit from the main `/Views` folder.

This can be mitigated by moving the `_ViewStart.cshtml` and `_ViewImports.chsml` to the root of the site. While I don't like cluttering up the root with more files, I do appreciate that this allows us to define our `@using`, `@addTagHelper`, and `Layout` declarations in one place.

Based on this, I was able to either trim down or entirely remove most of the `_ViewStart.cshtml` and `_ViewImports.cshtml` files from within `/Areas`, since much of what they defined was duplicating the main `/Views` copy.