Established SSL for local
As of Chrome 84, released on July 28th, all cookies default to `samesite=lax`. This breaks the AzureAD OpenID Connect Single Sign On authentication. This can be fixed by explicitly setting the `CorrelationCookies.SameSite` for OpenID to `SameSiteMode.None`.

For Chrome, however, this also requires use of SSL—which is already enforced in our staging and www environments, but wasn't being enforced in `localhost`. As such, this update is also being paired with an update to `launchSettings.json` to enable HTTPS for both IIS Express as well as the Kestrel web server.