Moved explicit routes to top, reformatted defaults
As we prepare to move away from explicit routes in preference of broader patterns, I'm moving the two explicit routes we intend to keep to the top of the endpoint configuration, so they're grouped together.

In addition, I'm refromatting the `defaults` for consistency with our JSON formatting—though, admittedly, doing so makes them take up more room than may be necessary. We may want to reevaluate this later, based on internal review.