Introduce `MapDefaultControllerRoute()`
Previously, we explicitly mapped a `{controller}/{action=Index}` route. This is supiciously similar to the out-of-the-box `MapDefaultControllerRoute()`, which offers a bit more flexibility while satisfying that requirement. In addition, this _also_ satisfies the requirement of the `MapTopicRoute("Error", "Error")` call. As such, these two can be consolidated using `MapDefaultControllerRoute()`.

While I was at it, I also added comments to each line to provide a quick hint about what the route pattern is, since this is now made more opaque by the use of the otherwise more convenient extension methods.