Updated to OnTopic Editor 4.4.0
In addition to taking advantage of the updates to OnTopic 4.4.0—one of which has a dramatic impact on OnTopic Editor—the 4.4.0 release also includes updates to how relationship attributes are handled.

Multiple relationships can now be placed on a single topic without interfering with one another. In addition, relationships are now, by default, automatically expanded to show all selected nodes—though this feature can be disabled via a flag on the `RelationshipAttribute` descriptor.

Finally, this update includes various bug fixes and maintenance updates to dependencies.