Removed implicit controller hack from `FormsController`
When we moved the `FormsController` to a new `/Area/Forms`, we wanted the controller to be implicit, so you didn't need to route to e.g. `/Forms/Forms/Quote`. This worked fine in terms of routing, but ran into issues with the OnTopic library, however, as its `ITopicRepository.Load(RouteValues)` would process both `{area}` and `{controller}`. We hacked around this by explicitly removing the `{controller}` when any action on the `FormsController` was called.

With OnTopic 4.4.0, this hack is no longer necessary, as the `ITopicRepository.Load(RouteValues)` has been updated to account for implicit controllers where the `{area}` and the `{controller}` are the same.

Good riddance!