Removed `Customer` and `CustomerUrl` from `ExampleApplication`
GoldSim was only using the `Customer` attribute on two `ExampleApplication`s, and the `CustomerUrl` on one. Given that, Rick has opted to move these into the body content instead of having them listed in the header area.

To ensure the code reflects this change, I've removed the `Customer` and `CustomerUrl` properties from the `ExampleApplicationTopicViewModel`, as well as their references from the `ExampleApplication.cshtml`. This also allowed me to get rid of the `getCustomerLink()` helper function on the view, since it's no longer used.