Removed `DownloadUrl` from `ExampleApplication` template
While this looks nice, the original purpose for it is no longer relevant, and so it's no longer used. As part of this, I was able to remove some of the complexity around maintaining a grid just for the `PostTitleSection`. Instead, the metadata can be a simple `<p class="metadata" />`.