Introduced redirects for legacy files
Rick has been cleaning up the file and folder structure of the media server. For the most part, those files are only referenced via OnTopic and can be taken care of via a search and replace across the topic graph. A handful of files, however, are referenced from distributed documentation, if not the application itself, and thus need to maintain redirects.

Unfortunately, as their target location varies, these can't be handled via a single rule, and thus we're implementing them as `PageMoves`.