Changed order of `MapDefault*ControllerRoute()`
If the new `MapDefaultsAreaControllerRoute()` takes place before the out-of-the-box `MapDefaultControllerRoute()`, then the `@Url.Action()` helper will return e.g. `/Forms/Forms/Quote` instead of `/Forms/Quote`.

This _should_ be getting mapped to `MapImplicitAreaControllerRoute()` but, due to a bug in ASP.NET Core, that doesn't work. As such, it makes sense that it would get picked up by `MapDefaultAreaControllerRoute()`. Once that bug is fixed—hopefully in ASP.NET 5.0—the order here won't matter as much.

Until then, we can hack around this by flipping the order of these. This is a hack because it means that area routes are getting intercepted by the more general `MapDefaultControllerRoute()` first, which simply excludes the `{area}` entirely, at least as far as `@Url.Action()` is concerned.

But this prevents us from needing to manually register `Courses` and `Forms` since they both rely on the `MapImplicitAreaControllerRoute()`.