Reestablished explicit administration route
While the hack introduced in 1a797b6 prevents the need to explicitly register the `Courses` and `Forms` routes, it unfortunately introduces the need to explicitly register the `Administration` route, as we previously did. That's still simpler than needing to register both, but we'll definitely want to reevaluate this in ASP.NET 5.0.