Introduced `IncludeNewsletter` and `IncludeWebinar` options to Newsletter
Jason would like users to be able to subscribe to the webinar mailing list at the same time as they sign up for the main newsletter. To make this clear, I've added a list of newsletters that the user can subscribe to, with the main newsletter being checked by default. That way, the user can opt-out of the main newsletter while opting into the webinar newsletter, if they choose.