Fixed path to Application Areas images
While most of the images on the homepage point to images that are part of this repository, the home page shares graphic assets with the modules pages, and thus maintained a hard-coded reference to the legacy `/Images/Content/` directory. This has been updated to the new `//media.GoldSim.com/Images/Web/` directory.