Fixed paths to module icons
While most of the images in the stylesheet point to files that are part of this repository, the Product landing page shares icon assets with the modules pages, and thus maintained a hard-coded reference to the legacy `/Images/Content/` directory. This has been updated to the new `//media.GoldSim.com/Images/Web/` directory.