Refactored white papers as `WhitePaperList`
In a previous commit, I refactored the white papers as a standard `ContentList`. That works fine. But it doesn't allow the editor to provide an association with the application areas, since `ContentItem` doesn't otherwise require that relationship. To permit this, I subclassed the `ContentList` and `ContentItem` content types as `WhitePaperList` and `WhitePaper`, and added an `Applications` relationship to `WhitePaper`.

To maintain compatibility with this, I introduced a new `WhitePaperListTopicViewModel`, `WhitePaperTopicViewModel`, and `WhitePaperList.cshtml`. Likewise, the `ApplicationPageTopicViewModel` now provides a strongly typed collection of `WhitePaperTopicViewModel` instances in its `WhitePapers` property. These each derive most of their functionality from other classes, and are thus trivial additions. In doing so, however, they afford us a lot more flexibility in the future.