Registered `PostmarkSmtpService` dependency
In order to share SMTP clients, while also gaining access to the API key via configuration, the `PostmarkSmtpService` is configured using the singleton lifestyle in `GoldSimActivator`.

I've commented out the legacy `SendGridApiKey` for now, just in case we need to return to it for some reason; that offers a backup in case we run into any issues during the migration.