Removed legacy compile exceptions
Previously, we had a number of directories that were not intended to be compiled or deployed. These included content directories, such as `Downloads` and `Media`, as well as legacy ASP.NET Web Forms directories, such as `!Admin` (for the OnTopic Editor), `Common` (for shared code), and `Forms` (which had not been migrated to ASP.NET MVC Framework in the 2017 redesign). Those are all removed now, and no longer need to be carved out as exceptions.