Added redirects for top four 404s
While these aren't critical, they consistently top the 404 charts and get a surprising amount of organic traffic from search engines. A couple of these probably don't matter, but will at least quiet the noise. The other two are relevant to the product and may be useful to maintain the page rank.