Updated `canonical` metadata to use `WebPath`
The `WebPath` is a true canonical in that it maps to the exact path and casing of the topic graph. This is better than using the `Context.Request.Path` which will simply use whatever casing is in the path. E.g., for `/Web/Products/`, this might be `/web/products/`, `/Web/Products`, or even `/wEb/PrOdUcTs//`. That doesn't matter as far as the web server is concerned, but these are each distinct URLs from Google's perspective.