Replace cookie consent animation with `SlideDown()`
In a previous commit, I introduced functionality that allowed us to show (9a81c75) and hide (9a96038) the `#CookiesNotice` element with an animation. This built on top of the existing CSS, which required some workarounds to ensure that the animation worked since, by default, a CSS `transform` doesn't animate a transition from `display: none`.

In this update, I replace the original implementation entirely with jQuery's `SlideDown()` and `SlideUp()` functions. This eliminates the need for the workarounds, and allows us to entirely remove the legacy `has-notice` class. This maintains the same benefits in terms of Cumulative Layout Shift (CLS), while reducing the complexity of the CSS. (It also reduces the JavaScript size by a small amount.)