Remove legacy header animation
Similar to the new `#CookiesNotice` animation (b87be48), the header used to have an animation associated with it which permitted it to slide up and down. This was in response to user scrolling, so the header would disappear as the user scrolled down, and reappear as the user scrolled up—functionality that was aided by the third-party Headroom.js library.

There were problems with this implementation, and it was never maintained on the production site. As part of my JavaScript cleanup a couple of months ago, I removed Headroom (1e81bed). When I did this, however, I didn't remove the related custom CSS associated with the implementation.

At some point, I'd like to revisit this. It was a handy feature, and especially given how long some of the pages are. The implementation would likely be different, however, since the libraries have likely evolved and matured, and our original implementation was buggy. Given that, we'll just remove these styles for now.